refactor to utility function
